Fix stupid client/client thing
