+++ lib/cortex-client.rb
@@ -1,2 +1,2 @@
-require 'cortex/client/client'
+require 'cortex/client'
 require 'cortex/client/version'
