+++ lib/cortex/client/version.rb
@@ -1,5 +1,5 @@
 module Cortex
-  module Client
+  class Client
     VERSION = '0.4.0'
   end
 end
